+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ### HEAD
 
+- simplify library entities:
+  - `SagroneScraper::Parser` is now `SagroneScraper::Base`
+  - `SagroneScraper::Base` is base class to create new scrapers
+  - `SagroneScraper.registered_parsers` is now `SagroneScraper.registered_scrapers`
+  - `SagroneScraper.register_parser` is now `SagroneScraper.register_scraper`
+
 ### 0.0.3
 
 - add `SagroneScraper::Parser.can_parse?(url)` class method, which must be  implemented in subclasses